requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
pandas==3.0.1
propcache==0.4.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torch-geometric==2.7.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
